+++ tests/test_reader.py
@@ -436,6 +436,17 @@ def test_mark_as_read_unread(reader, entry_arg):
     0,                                  # unchunked query
 ])
 def test_get_entries_order(reader, chunk_size):
+    """Entries should be sorted descending by (with decreasing priority):
+
+    * entry published
+    * entry updated
+    * feed URL
+    * order of entry in feed
+    * entry id
+
+    https://github.com/lemon24/reader/issues/97
+
+    """
     reader._get_entries_chunk_size = chunk_size
 
     parser = Parser()
@@ -445,7 +456,7 @@ def test_get_entries_order(reader, chunk_size):
     two = parser.feed(2)
     reader.add_feed(two.url)
 
-    parser.entry(2, 1, datetime(2010, 1, 1))
+    parser.entry(2, 1, datetime(2010, 1, 1), published=datetime(2010, 1, 1))
     parser.entry(2, 4, datetime(2010, 1, 4))
     two = parser.feed(2, datetime(2010, 1, 4))
     reader.update_feeds()
@@ -466,14 +477,14 @@ def test_get_entries_order(reader, chunk_size):
     parser.entry(1, 3, datetime(2010, 1, 4))
     one = parser.feed(1, datetime(2010, 1, 6))
     parser.entry(2, 3, datetime(2010, 1, 2))
-    parser.entry(2, 5, datetime(2010, 1, 3))
+    parser.entry(2, 5, datetime(2010, 1, 3), published=datetime(2009, 12, 20))
     two = parser.feed(2, datetime(2010, 1, 6))
     reader.update_feeds()
 
-    expected = sorted(
-        parser.get_tuples(),
-        key=lambda e: (e.updated, e.feed.url, e.id),
-        reverse=True)
+    def sort_key(e):
+        return e.published or e.updated, e.feed.url, e.id
+
+    expected = sorted(parser.get_tuples(), key=sort_key, reverse=True)
 
     assert list(reader.get_entries()) == expected
 
@@ -486,6 +497,8 @@ def test_get_entries_order(reader, chunk_size):
 def test_get_entries_feed_order(reader, chunk_size):
     """All other things being equal, get_entries() should yield entries
     in the order they appear in the feed.
+
+    https://github.com/lemon24/reader/issues/87
 
     """
     reader._get_entries_chunk_size = chunk_size